feat: slug prefixes /e/ and /p/ + centralized route map (#232)
- Move event pages: [eventId] → e/[eventId]
- Move profile pages: profile/[handle] → profile/p/[handle]
- Add redirect pages at old paths to preserve backwards compatibility
- Create comprehensive route map in packages/config/src/routes.ts
- Update all internal links in events, kernel, and shared libs
- Type check passes (pre-existing errors unchanged)

closes #232
