+++ apps/events/app/api/events/[id]/invites/route.ts
@@ -9,6 +9,7 @@ import { eq } from 'drizzle-orm';
 import { requireAuth } from '@imajin/auth';
 import { isEventOrganizer } from '@/src/lib/organizer';
 import { randomBytes } from 'crypto';
+import { eventUrl } from '@imajin/config';
 
 const EVENTS_URL = process.env.NEXT_PUBLIC_EVENTS_URL || 'https://events.imajin.ai';
 
@@ -35,7 +36,7 @@ export async function GET(
 
   const withUrls = invites.map(inv => ({
     ...inv,
-    url: `${EVENTS_URL}/${id}?invite=${inv.token}`,
+    url: `${eventUrl(EVENTS_URL, id)}?invite=${inv.token}`,
   }));
 
   return NextResponse.json({ invites: withUrls });
@@ -78,7 +79,7 @@ export async function POST(
   return NextResponse.json({
     invite: {
       ...invite,
-      url: `${EVENTS_URL}/${id}?invite=${invite.token}`,
+      url: `${eventUrl(EVENTS_URL, id)}?invite=${invite.token}`,
     },
   });
 }
