+++ apps/events/app/e/[eventId]/components/EventChatWrapper.tsx
