+++ apps/events/app/e/[eventId]/event-lobby-accordion.tsx
