+++ apps/events/app/e/[eventId]/magic-link-button.tsx
