+++ apps/events/app/e/[eventId]/tickets-section.tsx
